06-deepagent/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.84.0
anyio==4.12.1
asttokens==3.0.1
attrs==25.4.0
blockbuster==1.5.26
bracex==2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
comm==0.2.3
croniter==6.0.0
cryptography==46.0.5
daytona==0.149.0
daytona-api-client==0.149.0
daytona-api-client-async==0.149.0
daytona-toolbox-api-client==0.149.0
daytona-toolbox-api-client-async==0.149.0
debugpy==1.8.20
decorator==5.2.1
deepagents==0.4.7
deprecated==1.3.1
distro==1.9.0
docstring-parser==0.17.0
environs==14.6.0
executing==2.2.1
filetype==1.2.0
forbiddenfruit==0.1.4
frozenlist==1.8.0
google-auth==2.49.0
google-genai==1.66.0
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-health-checking==1.78.0
grpcio-tools==1.78.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib-metadata==8.7.1
ipykernel==7.2.0
ipython==9.11.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema-rs==0.44.1
jupyter-client==8.8.0
jupyter-core==5.9.1
langchain==1.2.10
langchain-anthropic==1.3.4
langchain-core==1.2.18
langchain-daytona==0.0.3
langchain-google-genai==4.2.1
langchain-openai==1.1.11
langgraph==1.0.10
langgraph-api==0.7.68
langgraph-checkpoint==4.0.1
langgraph-cli==0.4.15
langgraph-prebuilt==1.0.8
langgraph-runtime-inmem==0.26.0
langgraph-sdk==0.3.10
langsmith==0.7.16
marshmallow==4.2.2
matplotlib-inline==0.2.1
multidict==6.7.1
multipart==1.3.1
nest-asyncio==1.6.0
obstore==0.8.2
openai==2.26.0
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-http==1.40.0
opentelemetry-instrumentation==0.61b0
opentelemetry-instrumentation-aiohttp-client==0.61b0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
opentelemetry-util-http==0.61b0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
parso==0.8.6
platformdirs==4.9.4
prompt-toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
pure-eval==0.2.3
pyasn1==0.6.2
pyasn1-modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pygments==2.19.2
pyjwt==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
pyyaml==6.0.3
pyzmq==27.1.0
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
setuptools==82.0.1
six==1.17.0
slack-sdk==3.40.1
sniffio==1.3.1
sse-starlette==3.3.2
stack-data==0.6.3
starlette==0.52.1
structlog==25.5.0
tavily-python==0.7.23
tenacity==9.1.4
tiktoken==0.12.0
toml==0.10.2
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
truststore==0.10.4
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uuid-utils==0.14.1
uvicorn==0.41.0
watchfiles==1.1.1
wcmatch==10.1
wcwidth==0.6.0
websockets==15.0.1
wrapt==1.17.3
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
zstandard==0.25.0
